support for windows paths
separate windows and unix tests, close file handles

fix linting errors

make TestBasicTail a universal test again

remove O_APPEND from OpenLogFile to support truncating on Windows

Add further documentation why tests are unix-specific

fix wrong octal notation

fix linting errors
