+++ internal/mtail/compile_only_integration_test.go
@@ -19,7 +19,7 @@ func TestBadProgramFailsCompilation(t *testing.T) {
 	testutil.SkipIfShort(t)
 	progDir := testutil.TestTempDir(t)
 
-	err := ioutil.WriteFile(filepath.Join(progDir, "bad.mtail"), []byte("asdfasdf\n"), 0666)
+	err := ioutil.WriteFile(filepath.Join(progDir, "bad.mtail"), []byte("asdfasdf\n"), 0o666)
 	testutil.FatalIfErr(t, err)
 
 	ctx := context.Background()
