+++ internal/runtime/fuzz.go
@@ -1,6 +1,7 @@
 // Copyright 2011 Google Inc. All Rights Reserved.
 // This file is available under the Apache license.
 
+//go:build gofuzz
 // +build gofuzz
 
 package runtime
