+++ internal/tailer/logstream/pipestream.go
@@ -43,7 +43,7 @@ func (ps *pipeStream) LastReadTime() time.Time {
 
 func (ps *pipeStream) stream(ctx context.Context, wg *sync.WaitGroup, waker waker.Waker, fi os.FileInfo) error {
 	// Open in nonblocking mode because the write end of the pipe may not have started yet.
-	fd, err := os.OpenFile(ps.pathname, os.O_RDONLY|syscall.O_NONBLOCK, 0600)
+	fd, err := os.OpenFile(ps.pathname, os.O_RDONLY|syscall.O_NONBLOCK, 0o600)
 	if err != nil {
 		logErrors.Add(ps.pathname, 1)
 		return err
