+++ internal/testutil/race.go
@@ -1,5 +1,6 @@
 // Copyright 2021 Google Inc. All Rights Reserved.
 // This file is available under the Apache license.
+//go:build race
 // +build race
 
 package testutil
